first version with parameters in command line first version with check file difference using diff

+++ syncDirectories.py
@@ -3,32 +3,65 @@
 
 # Synchronize directory <SOURCE_DIR>/<SUB_DIR> with directory <DEST_DIR>/<SUB_DIR>
 #
-# Requires Python 3
+# Requires:
+# - Python 3
+# - "diff" directory in this script's current directory
 
 
 import os
+import platform
+import shlex
 import shutil
+import sys, getopt
+
+from subprocess import Popen, PIPE
 
 
-def main():
+def main(argv):
 
-    # Directories
-    SUB_DIR    = ""
-    SOURCE_DIR = "C:\Dev\SyncDirectories\SRC"
-    DEST_DIR   = "C:\Dev\SyncDirectories\DEST"
+    SOURCE_DIR = ""
+    DEST_DIR = ""
+    checkFilesDifference = False
+    simulate = False
 
-    # Add SUB_DIR to source and destination directories
-    if SUB_DIR:
-        SOURCE_DIR = os.path.join(SOURCE_DIR, SUB_DIR)
-        DEST_DIR   = os.path.join(DEST_DIR, SUB_DIR)
+    try:
+        opts, args = getopt.getopt(argv,"hs:d:ct",["help","srcdir=","dstdir=","checkdiff","test"])
+    except getopt.GetoptError:
+        printHelp()
+        sys.exit(2)
+    
+    for opt, arg in opts:
+        if opt in ("-h", "--help"):
+            printHelp()
+            sys.exit()
+        elif opt in ("-s", "--srcdir"):
+            SOURCE_DIR = arg
+        elif opt in ("-d", "--dstdir"):
+            DEST_DIR = arg
+        elif opt == '-c':
+            # Check precise files differences
+            print("Check files differences ON")
+            checkFilesDifference = True
+        elif opt == '-t':
+            # Test simulate synchronization
+            print("Simulation test ON")   
+            simulate = True
+   
+    print('Source directory:      "', SOURCE_DIR, '"')
+    print('Destination directory: "', DEST_DIR, '"')
+    print()
+    
+    if(not SOURCE_DIR or not DEST_DIR):
+        print('Error: source or destination directory is missing')
+        return
 
-    syncDirectoryClean(SOURCE_DIR, DEST_DIR)
-    syncDirectoryCopy(SOURCE_DIR, DEST_DIR)
-
+    # Synchronize directories
+    syncDirectoryClean(SOURCE_DIR, DEST_DIR, simulate)
+    syncDirectoryCopy(SOURCE_DIR, DEST_DIR, checkFilesDifference, simulate)
     print("Directories have been synchronized")
 
 
-def syncDirectoryClean(SOURCE_DIR, DEST_DIR):
+def syncDirectoryClean(SOURCE_DIR, DEST_DIR, simulate = False):
 
     # Add "/" to source and destination directories if necessary
     if(SOURCE_DIR[len(SOURCE_DIR) - 1] != "/"):
@@ -45,11 +78,13 @@ def syncDirectoryClean(SOURCE_DIR, DEST_DIR):
     # Create destination directory if not existing
     if(not os.path.isdir(DEST_DIR)):
         print("Creating directory", DEST_DIR)
-        os.makedirs(DEST_DIR)
-        
-    os.chdir(DEST_DIR)
-    print("CLEANING", SOURCE_DIR, "->", DEST_DIR)
-    print()
+        if(not simulate):
+            os.makedirs(DEST_DIR)
+
+    if(not simulate):  
+        os.chdir(DEST_DIR)
+        print("CLEANING", SOURCE_DIR, "->", DEST_DIR)
+        print()
 
     # Rename source files to destination directory
     nbTotal  = 0
@@ -63,7 +98,8 @@ def syncDirectoryClean(SOURCE_DIR, DEST_DIR):
         srcDir = os.path.join(SOURCE_DIR, localDir)
         if(not os.path.isdir(srcDir)):
             print("Removing directory", dirname)
-            shutil.rmtree(dirname)
+            if(not simulate):
+                shutil.rmtree(dirname)
             continue
 
         # Browse files
@@ -77,7 +113,8 @@ def syncDirectoryClean(SOURCE_DIR, DEST_DIR):
             if(not os.path.exists(srcFile)):
                 nbRemoved += 1
                 print("Removing", dstFile)
-                os.remove(dstFile)
+                if(not simulate):
+                    os.remove(dstFile)
 
 
     # Print result
@@ -87,7 +124,10 @@ def syncDirectoryClean(SOURCE_DIR, DEST_DIR):
     print()
 
 
-def syncDirectoryCopy(SOURCE_DIR, DEST_DIR):
+def syncDirectoryCopy(SOURCE_DIR, DEST_DIR, checkFilesDifference = False, simulate = False):
+
+    # Get program directory
+    programDir = os.path.dirname(os.path.realpath(__file__))
 
     # Add "/" to source and destination directories if necessary
     if(SOURCE_DIR[len(SOURCE_DIR) - 1] != "/"):
@@ -122,7 +162,8 @@ def syncDirectoryCopy(SOURCE_DIR, DEST_DIR):
             tmpDir = os.path.join(DEST_DIR, localDir)
             if(not os.path.isdir(tmpDir)):
                 print("Creating directory", tmpDir)
-                os.makedirs(tmpDir)
+                if(not simulate):
+                    os.makedirs(tmpDir)
 
             # If file does not exist in destination directory, copy it
             srcFile = os.path.join(dirname, filename)
@@ -132,27 +173,17 @@ def syncDirectoryCopy(SOURCE_DIR, DEST_DIR):
                 # Copy file
                 nbCopied += 1
                 print("Copying", srcFile)
-                shutil.copyfile(srcFile, dstFile)
+                if(not simulate):
+                    shutil.copyfile(srcFile, dstFile)
             else:
-                # Check sizes
-                srcSize = os.path.getsize(srcFile)
-                dstSize = os.path.getsize(dstFile)
-                differentSizes = (srcSize != dstSize)
-                #print("Sizes = ", srcSize, dstSize, differentSizes)
+                replaceFile = areFilesDifferent(srcFile, dstFile, checkFilesDifference, programDir)
 
-                # Check last modification dates
-                # srcLastModifTime = os.path.getmtime(srcFile)
-                # dstLastModifTime = os.path.getmtime(dstFile)
-                #srcModified = (srcLastModifTime > dstLastModifTime)
-
-                replaceFile = differentSizes
-                #replaceFile = differentSizes or modified
-
-                # Replace file iff neded
+                # Replace file iff needed
                 if(replaceFile):
                     nbReplaced += 1
                     print("Replacing", srcFile)
-                    shutil.copyfile(srcFile, dstFile)
+                    if(not simulate):
+                        shutil.copyfile(srcFile, dstFile)
 
 
     # Print result
@@ -163,7 +194,57 @@ def syncDirectoryCopy(SOURCE_DIR, DEST_DIR):
     print()
 
 
-main()
+def areFilesDifferent(file1Path, file2Path, checkDifference = False, programDir = ""):
+    
+    # Check sizes
+    file1Size = os.path.getsize(file1Path)
+    file2Size = os.path.getsize(file2Path)
+    differentSizes = (file1Size != file2Size)
+    if(differentSizes):
+        return True    
 
-# Wait for user input to close program (Windows)
-os.system("pause")
+    # Disabled: Check last modification dates
+    # file1LastModifTime = os.path.getmtime(file1Path)
+    # file2LastModifTime = os.path.getmtime(file2Path)
+    #modified = (file2LastModifTime > file1LastModifTime)
+    #if(modified):
+    #    return True  
+
+    # Check files differences if option set
+    if(checkDifference):
+        file1Path = file1Path.replace('\\', '/')
+        file2Path = file2Path.replace('\\', '/')
+        if(not programDir):
+            programDir = os.path.dirname(os.path.realpath(__file__))
+        platformName = platform.system()
+        progFile = "diff"
+        if(platformName.lower() == "windows"):
+            progFile = "diff.exe"
+        diffFilePath = os.path.join(programDir, "diff", progFile).replace('\\', '/')
+        cmd = diffFilePath + ' "' + file1Path + '" "' + file2Path + '"'
+        #print("Diff command = " + cmd)
+        process = Popen(shlex.split(cmd), stdout=PIPE)
+        process.communicate()
+        exitCode = process.wait()
+        #print("Diff exit code = " + str(exitCode))
+        return (exitCode != 0)
+    else:
+        return False
+
+def printHelp():
+    print()
+    print('Usage: python syncDirectories.py -s <source_directory> -d <destination_directory>')
+    print()
+    print("Options:")
+    print("  -s, --srcdir <dir>:    source directory used as reference for synchronization")
+    print("  -d, --dstdir <dir>:    destination directory to synchronize")
+    print("  -c, --checkdiff:       check precise files difference (slower)")
+    print("  -t, --test:            test synchronization simulation (traces only, no effects)")
+    print("  -h, --help:            display help")
+
+
+if __name__ == "__main__":
+    main(sys.argv[1:])
+
+    # Wait for user input to close program (Windows)
+    os.system("pause")
